my_project/requirements.txt
Django==4.2.2
django-cors-headers==4.1.0
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
django-apscheduler==0.6.2
django-sslserver==0.22
psycopg2-binary==2.8.6
psycopg2==2.9.9
Pillow==9.3.2
xlrd==1.2.0
XlsxWriter==3.0.3
requests==2.25.0
scipy==1.9.0
numpy==1.23.2
loguru==0.6.0
pandas==1.1.4
python-docx==0.8.10
openpyxl==3.0.5
comtypes==1.1.14
fitz==0.0.1.dev2
PyMuPDF==1.19.0
frontend==0.0.3
urllib3==1.26.2
pycryptodome==3.17
license_authorize==1.0.4
vgis-utils==1.0.10
vgis-office==1.0.2
vgis-log==1.0.5
vgis-database==1.0.2
cryptography==41.0.1
toollib==1.2.5
waitress==2.1.2
djangorestframework-gis==1.0
channels==3.0.5
channels-redis==4.1.0
daphne==3.0.2
django-redis==5.4.0
drf-yasg==1.21.7
vgis_encrption==1.0.3
